+++ src/Component/search/EntitySimple.js
@@ -1,12 +1,12 @@
 import PropTypes from 'prop-types';
-import {Button} from 'reactstrap';
+import { Button } from 'reactstrap';
 import './SearchForm.css';
-import {getClickableEntityLink} from "../../Utils";
+import { getClickableEntityLink } from "../../Utils";
 
 /**
- * Highlight a descriptor with link to the reference vocabulary
+ * Highlight a entities with link to the reference vocabulary
  */
-const EntityDescriptorSimple = (props) => {
+const EntitySimple = (props) => {
     const {
         id,
         label,
@@ -16,7 +16,7 @@ const EntityDescriptorSimple = (props) => {
 
     return (
         <span className="entity">
-            <Button id={id} type="button" className="btn highlight-descriptor">
+            <Button id={id} type="button" className="btn highlight-entities">
                 <a className="entity-link" href={getClickableEntityLink(link)} target="_external_entity">
                     {label}
                 </a>
@@ -25,10 +25,10 @@ const EntityDescriptorSimple = (props) => {
     );
 }
 
-EntityDescriptorSimple.propTypes = {
+EntitySimple.propTypes = {
     id: PropTypes.number.isRequired,
     label: PropTypes.string.isRequired,
     link: PropTypes.string.isRequired
 }
 
-export default EntityDescriptorSimple;
+export default EntitySimple;