+++ src/Component/search/SearchResult.js
@@ -4,7 +4,7 @@ import { Link } from "react-router-dom";
 import { Col, Row } from "react-bootstrap";
 import { BiDownload } from "react-icons/bi";
 import { HiOutlineDocumentMagnifyingGlass } from "react-icons/hi2";
-import EntityDescriptorSimple from "./EntityDescriptorSimple";
+import EntitySimple from "./EntitySimple";
 import './SearchResult.css';
 
 /**
@@ -57,7 +57,7 @@ const SearchResult = (props) => {
         if (matchedEntities !== undefined) {
             matchedEntities.forEach((_e, _id) => {
                 _entities.push(
-                    <EntityDescriptorSimple
+                    <EntitySimple
                         key={_id}
                         id={_id}
                         label={_e.entityLabel}